Add trait for locale factories
There is no one way of obtaining locale data and doing locale operations:

 - On Linux functions working with the locale_t struct obtained from
   newlocale() function is the right way. They are sufficiently flexible and
   the results should be consistent with most applications.

 - But KDE has it's own additional settings. The Numeric and Monetary
   categories can be even set with parameters that don't actually match any
   existing locale. Gnome may have similar settings; I did not check yet. So
   we may want to read these _in addition to_ the glibc locale functions.

 - Windows do have most of the standard C library functions, except the
   unfortunately important nl_langinfo, but they are _not_ what is normally
   used there, they don't support UTF-8 there and they use non-standard
   locale identifiers. Instead we should use the separate native api from
   winnls.h.

 - I don't have access to MacOS, but I expect there will be some
   difference there.

 - Android does not have locale data installed on the C side at all. The only
   way to implement localization there is to take the data from CLDR and pack
   them in the application or to somehow extract them from the Java side
   which has them (from CLDR).

 - The CLDR data are more detailed than what we can extract from either
   system, so I expect some applications will want to use them anyway.

My current plan is to create a Locale structure, that will contain boxed
instances of individual components of the locale: CType, Numeric, Time,
Collate, Monetary, Messages, Paper, Name, Address, Telephone, Measurement and
ability to attach user-defined extensions that will be available in
user-defined formatting functions.

To simplify construction of the Locale class, it's constructor will take
a factory that will define appropriate sources of the information. It will be
possible to create factories that only define specific locale categories and
combine them (using the CompositeLocaleFactory) with other factories to
provide the rest.

This is the first step defining the base LocaleFactory trait.