Bindings for GNU LibC locale module on Linux.
On Linux the libc locale are standard method of providing locale information
and thanks to the nl_langinfo_l function all the necessary data is readily
available in a thread-safe way. We also bind iconv, which is part of libc on
Linux and can thus be conveniently used to convert from the locale charset to
Rust's internal utf-8 as we can't guarantee that user's locale is utf-8.