Move the original functions for getting locale data to macos module.
They were originally written for MacOS. On Linux the locale data format is
different and they wouldn't work there, but because I don't have MacOS to
test on and because they don't contain anything that wouldn't compile on
Linux, I am leaving them compiled in unconditionally so I can test they
compile.

The methods in Numeric and Time were deprecated.