Hide the linux and macos submodules
They are implementation detail and for locale it will be rather important to
maintain compatibility, so the public surface should be kept small.

If somebody wanted the bindings to libc locales, they can be extracted to
separate crate (suggested name: `langinfo`; it mainly binds `langinfo.h`),
but I don't see much need at the moment.