Basic implementation of the facet factory and cache.
Something that should be workable, so we can proceed.

I was thinking that perhaps instead of reference-counting each facet
separately we could somehow lock the whole `row` for the `Locale` and hand
out plain old references with appropriate lifetime, but on the other hand it
would complicate the interface as the `row` would have to become
user-visible.

Of course in the current state it looks like we could just `unsafe` away the
lifetime and hand out `&'static` references, because the cache entries are
never removed, but plan is to add support for overriding the factories and
then the cache will be cleared when setting them.