Pythony pattern for numbers, parsing
Parser is based on nom. It is unfortunately somewhat ugly because there
does not seem to be a way to indicate the input is complete, so the parser
has to be sprinkled with complete! wrappers rather liberally.

Another sources of ugly are that nom is not really usable for &str, so the
parser has to be &[u8] and cast back for the grapheme parser, and that
despite working on binary, many elementary parsers return completely
inappropriate char.

I have also considered combine, but defining a new elementary parser
there is rather complicated (I need to define the grapheme parser
whatever I use) and there does not seem to be any how-to for it.