Refactor extract_cldr for explicit order.
The point is that I want to override some values in the root locale (in
CLDR, the root locale is very Americanish and not really invariant), but
need the inherited values in any other locale to remain the original
ones. This change will allow loading the data first, overriding some,
and then serializing them.

* The new items module reads definition of the enum Item and tries to
  replicate its ordering in Python.
* The Locale class now reads the data in one phase and serializes them
  in a separate one.
* The Locale item map is initially copied from the parent, so if the
  parent value is later overridden, the override won't propagate.