Drop command-line options from extract_cldr script
As I intend to generate more output files, having options on
command-line no longer sounds sensible. The outputs are normally fixed
anyway and at each step, the best way to verify the output is as desired
is to compare the files in git. So always doing exactly the one intended
thing from now on.